refactor: use Interlocked for Probe ActiveCount
Simplify thread-safe increment and decrement Probe ActiveCount by using Interlocked rather than a mutex.

ActiveCount is changed to a long to accomodate this change. The related WPF value converters have been updated to expect long values.

+++ vmPing/Classes/ValueConverters.cs
@@ -336,7 +336,7 @@ namespace vmPing.Classes
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            if ((int)value > 0)
+            if ((long)value > 0)
             {
                 return Strings.Toolbar_StopAll;
             }
@@ -356,7 +356,7 @@ namespace vmPing.Classes
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            if ((int)value > 0)
+            if ((long)value > 0)
             {
                 return (DrawingImage)Application.Current.Resources["icon.stop-circle"];
             }
